[Refactor][2/N][Torchax] Move vllm layers
