+++ tests/layers/vllm/test_pallas_torchax.py
@@ -11,8 +11,8 @@ from torchax.interop import torch_view
 from vllm.attention.backends.abstract import AttentionType
 
 from tpu_inference.layers.common.attention_metadata import AttentionMetadata
-from tpu_inference.models.vllm.pallas_torchax import (PallasAttentionBackend,
-                                                    PallasAttentionBackendImpl)
+from tpu_inference.layers.vllm.pallas_torchax import (
+    PallasAttentionBackend, PallasAttentionBackendImpl)
 from tpu_inference.models.vllm.vllm_model_wrapper_context import \
     set_vllm_model_wrapper_context
 from tpu_inference.runner.kv_cache import get_kv_cache_shape_with_mesh
